update commands for categorizations
