+++ src/domain_manager/main.py
@@ -3,6 +3,7 @@
 import click
 from colorama import Fore
 from scripts.active_sites import active_sites
+from scripts.categorization import categories
 from scripts.dns_record_handler import hosted_zones
 from scripts.get_data import get_data
 
@@ -27,6 +28,7 @@ if __name__ == "__main__":
     cli.add_command(get_data)
     cli.add_command(hosted_zones)
     cli.add_command(active_sites)
+    cli.add_command(categories)
 
     # Run the command line application
     cli()
