+++ src/domain_manager/scripts/categorization.py
@@ -1,9 +1,16 @@
-"""A sample python script."""
+"""Check a domain's categories."""
 # Third-Party Libraries
+import click
 from colorama import Fore
 import requests
 from utils.message_handling import error_msg, success_msg
 from utils.settings import URL, auth
+
+
+@click.group()
+def categories():
+    """Manage proxy categorizations of active websites."""
+    pass
 
 
 def get_live_sites():
@@ -12,12 +19,16 @@ def get_live_sites():
     return resp.json()
 
 
+@categories.command("list")
 def get_categories():
     """Returns a list of categories."""
     resp = requests.get(f"{URL}/api/categories/", headers=auth)
+    category_names = "\n".join(category["name"] for category in resp.json())
+    success_msg(category_names)
     return resp.json()
 
 
+@categories.command("categorize")
 def categorize_live_site():
     """
     Categorize an active site.
@@ -30,11 +41,11 @@ def categorize_live_site():
 
     categories = [category.get("name") for category in get_categories()]
 
-    print(Fore.LIGHTBLUE_EX + "\n".join(categories) + Fore.WHITE)
+    click.echo(Fore.LIGHTBLUE_EX + "\n".join(categories) + Fore.WHITE)
 
     # Choose a category
     category_name = input("Please enter a category: ")
-    print(category_name)
+
     # Access live site data by uuid
     live_site_id = "".join(
         site.get("_id")
@@ -60,3 +71,22 @@ def categorize_live_site():
         success_msg(resp_json["message"])
 
     return resp_json
+
+
+@categories.command("check")
+def check_categories():
+    """Check a domain's categories on multiple proxies."""
+    domain = input("Please enter a domain name: ")
+    click.echo("Checking categories...")
+    resp = requests.get(f"{URL}/api/check/?domain={domain}", headers=auth)
+    try:
+        resp.raise_for_status()
+    except requests.exceptions.HTTPError as e:
+        error_msg(str(e))
+        return
+
+    for key, value in resp.json().items():
+        click.echo(Fore.GREEN + key + ": " + Fore.WHITE)
+        if value is not None:
+            success_msg(value)
+    return resp.json()
